Rearrange and quick manual pass to clean up split lines
